requirements.txt
python-dotenv>=1.0.0
openai>=1.0.0
google-generativeai>=0.3.0
aiohttp>=3.8.0
httpx[socks]>=0.24.0 